+++ src/qml/qml/v4/qv4bindings.cpp
@@ -52,6 +52,7 @@
 #include <private/qjsconverter_p.h>
 #include <private/qjsconverter_impl_p.h>
 #include <private/qjsvalue_impl_p.h>
+#include <private/qjsvalueiterator_impl_p.h>
 #include <private/qv8engine_impl_p.h>
 
 #include <private/qqmlaccessors_p.h>
@@ -864,7 +865,7 @@ inline quint32 QV4Bindings::toUint32(double n)
         QQmlPropertyData *prop = (data && data->propertyCache) ? data->propertyCache->property((index)) : 0; \
         if (prop && prop->isOverridden()) { \
             int resolvedIndex = data->propertyCache->property(prop->name(obj))->coreIndex; \
-            if (index < resolvedIndex) { \
+            if ((int)index < resolvedIndex) { \
                 *(inv) = true; \
                 goto programExit; \
             } \
